summarization/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.4.0
dill==0.3.8
evaluate==0.4.3
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.3
idna==3.10
joblib==1.4.2
multidict==6.1.0
multiprocess==0.70.16
nltk==3.9.1
numpy==1.23.5
packaging==24.2
pandas==1.5.3
propcache==0.3.0
pyarrow==19.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
scikit-learn==1.2.2
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.67.1
transformers==4.29.2
typing_extensions==4.12.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3


# requirements.txt
# Project Dependencies for Documentation Summarization AI

# Core Machine Learning and NLP Libraries
torch==1.13.1
transformers==4.29.2
sentencepiece==0.1.97

# Data Manipulation and Scientific Computing
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2

# Configuration and Logging
PyYAML==6.0
logging==0.5.1.2

# Evaluation and Metrics
rouge-score==0.1.2
evaluate==0.4.0

# Development and Testing
pytest==7.3.1
pytest-cov==4.0.0

# Optional: Jupyter Notebook Support
jupyter==1.0.0
notebook==6.5.4

# Utilities
tqdm==4.65.0